requirements.txt
polygon-api-client
google-cloud-bigquery